requirements.txt
eventlet==0.30.2
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-JWT-Extended==3.25.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.6.0
flask-rest-paginate==1.1.2
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.2
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
gevent-websocket==0.10.1
gunicorn==20.1.0
json5==0.9.5
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
python-dotenv==0.15.0
python-socketio==4.6.1
requests==2.25.1
SQLAlchemy==1.4.3
spotipy==2.17.1
sklearn==0.0
Werkzeug==0.16.0
